refactor(cli): move the .NET toolchain probes out of doctor.ts
The .NET SDK and MAUI workload probes and the gate that offers to install the
workload move into dotnet-toolchain.ts, unchanged. The scaffolder runs that
gate, and importing it from doctor.ts pulls in the config loader (jiti, the
TypeScript compiler), signing and Velopack. That matters once the scaffolder
ships as its own package, and esbuild's tree-shaking does not remove those
imports.

Two helpers move to where they belong rather than into the new module: run()
joins exec.ts, where toText already did what doctor.ts's bufToStr duplicated,
and splitLines/isInteractive join utils.ts. No behaviour changes.
