+++ src/cli/create-vidra-app/src/dev-port.ts
@@ -1,6 +1,6 @@
 import { createServer } from "node:net";
 import prompts from "prompts";
-import { isInteractive } from "./doctor.js";
+import { isInteractive } from "./utils.js";
 
 const MIN_PORT = 1;
 const MAX_PORT = 65_535;
